requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
psycopg[binary]==3.1.13
Werkzeug==3.0.1
Flask-Mail==0.9.1
gunicorn==21.2.0